[expotools] Publish beta project templates
Why

Currently, we ask only if you want to tag as latest, and we do it for each template.

In create-expo-app we want to avoid having to be aware of the versions endpoint, which we currently use in EXPO_BETA=1 expo init to look for the SDK version with the beta field set. If we tag templates as beta on npm then we can make EXPO_BETA=1 npx create-expo-app initialize from template-name@beta.

How

Modify the flow as follows:

- User is asked the Expo SDK version for the templates (no change here)
- User is then presented a choice of tags for the release:
  - sdk-xx and latest
  - sdk-xx and beta
  - sdk-xx
  - custom
- If "custom" is selected above, user types in a custom tag to be used
- Then for each package selected:
  - User selects the new version for the package
  - Script executes "npm publish" with the new version and with tag sdk-xx or custom, per the choices the user made above
  - If the user has selected the right options, the script executes "npm dist-tag" with either the "latest" or "beta" tag

Test Plan

All options tested in dry-run testing locally.